(jvm) corrigindo ifcmpl -> offset é inteiro 16 bits e nao 32

+++ src/jvm/instr_control.c
@@ -152,7 +152,7 @@ void handle_ifcmp(JVM_Context* ctx, u1 opc)
 {
   Frame* frame = current_frame(ctx);
   u1* ifcond_pc = frame->pc - 1; // inicio do if
-  int32_t offset = (int32_t)fetch_u2(&frame->pc);
+  int16_t offset = (int16_t)fetch_u2(&frame->pc);
 
   int32_t v2 = (int32_t)pop_operand(frame);
   int32_t v1 = (int32_t)pop_operand(frame);
